- fixes compilation issues on 64bit Darwin.
git-svn-id: http://svn.berlios.de/svnroot/repos/sems/trunk@1711 8eb893ce-cfd4-0310-b710-fb5ebe64c474

+++ core/AmSession.cpp
@@ -425,7 +425,7 @@ string AmSession::getNewId()
 
   id += int2hex(get_random()) + "-";
   id += int2hex(t.tv_sec) + int2hex(t.tv_usec) + "-";
-  id += int2hex((unsigned int) pthread_self());
+  id += int2hex((unsigned int)((unsigned long)pthread_self()));
 
   return id;
 }
